+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Rewritten source code to Lua
+- msgpack.NULL is valid for optional parameters now (incompatible change)
+
+### Added
+- Ability to check "option" parameters following their hierarchical structure.
+- Tests
+
 ## [1.0.0] - 2018-02-25
 ### Added
 - Imported checks library from luarocks