+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0] - 2018-06-19
 ### Added
 - Built-in checkers: `int64`, `uint64`
 - Built-in checkers: `uuid`, `uuid_str`, `uuid_bin`